Rename async_request module to sql_statement_metadata
In merging the AsyncSqlClient and SqlClient protocols it became
apparent that the async_request module did not contain any references
to async behavior or async components. Rather, it provides some
classes aimed at supporting SQL query tagging behavior. This rename
better reflects the purpose of this module, as we plan to enable
optional tag support in the main query API and eventually remove
the async query methods.
