+++ metricflow/sql_clients/sqlalchemy_dialect.py
@@ -14,9 +14,9 @@ from metricflow.dataflow.sql_table import SqlTable
 from metricflow.protocols.sql_client import SqlIsolationLevel
 from metricflow.protocols.sql_request import SqlJsonTag, SqlRequestTagSet
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
-from metricflow.sql_clients.async_request import CombinedSqlTags
 from metricflow.sql_clients.base_sql_client_implementation import BaseSqlClientImplementation
 from metricflow.sql_clients.common_client import check_isolation_level
+from metricflow.sql_clients.sql_statement_metadata import CombinedSqlTags
 
 logger = logging.getLogger(__name__)
